pyproject.toml
[project]
name = "mcp-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastmcp>=2.13.1",
    "llama-cpp-python>=0.3.16",
    "mcp[cli]>=1.22.0",
    "playwright>=1.56.0",
    "pydantic>=2.12.5",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "typing>=3.10.0.0",
]
